Refactor StoryMainVC's navController usage to work better in SplitVC
It was intractable mess to figure out how to make pushing the StoryMainVC
work correctly on large iPhones which behave like phones in portrait but
iPads in landscape. Whether to use the SplitVC's primary or secondary
NavController to push the story VC depended on which SplitVC display
mode iOS has decided to be in, and if you rotate the device after
pushing on one NC, it would need to magically move to the other NC to be
consistent.

Simplified things greatly by always using presentViewController and an
independent NC just for the StoryMainVC.  Unfortunately there's no
presentation style for this that looks the same as if you were pushing
a view controller onto a navController (slide in from right), and
I didn't want to bother making a custom transition/animation, but
using cross-fade looks decent.