Handle zip downloads from IFDB better.
Use the 'Contains' marker to identify the story filename to associate
metadats with.  Add a dialog to pick which files to extract from zip
instead of doing all of them.  Auto-select only the story that matches
the IFDB page for the download.  Did the dialog in Swift, because why
not.